Deleting Files

+++ Section15/controllers/admin.js
@@ -2,6 +2,7 @@ const Product = require("../models/product");
 const { validationResult } = require('express-validator');
 
 const mongoose = require('mongoose')
+const fileHelper = require('../util/file.js')
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -162,6 +163,7 @@ exports.postEditProduct = (req, res, next) => {
     product.title = updatedTitle;
     product.price = updatedPrice;
     if(image){
+      fileHelper.deleteFile(product.imageUrl)
       product.imageUrl = image?.path;
     }
     product.description = updatedDesc;
@@ -181,18 +183,25 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      // console.log(result);
-      console.log("DELETED PRODUCT!!!");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => {
-      console.log(err);
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(prodId)
+  .then(product => {
+    if(!product){
+      return next(new Error('Product not found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({_id: prodId, userId: req.user._id})
+  }).then(() => {
+    // console.log(result);
+    console.log("DELETED PRODUCT!!!");
+    res.redirect("/admin/products");
+  })
+  .catch((err) => {
+    console.log(err);
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  });
+    
   // Product.deleteById(prodId)
   // res.redirect('/admin/products')
 };
